+++ src/codespaces/azext_codespaces/_params.py
@@ -19,14 +19,20 @@ def load_arguments(self, _):
         c.argument('plan_name', options_list=['--plan', '-p'], help="Name or ID of the Codespace plan", validator=validate_plan_name_or_id)
         c.argument('codespace_name', options_list=['--name', '-n'], help='Name of the Codespace.')
         c.argument('codespace_id', options_list=['--id'], help='Id of the Codespace.', validator=validate_codespace_name_or_id)
+        c.argument('sku_name', options_list=['--instance-type'], help='Instance Type')
+        c.argument('autoshutdown_delay', options_list=['--suspend-after'], arg_type=get_enum_type(['5', '30', '120']), help="Automatically suspend the inactive Codespace after this many minutes.")
 
     with self.argument_context('codespace plan') as c:
         c.argument('plan_name', options_list=['--name', '-n'], help="Name of the Codespace plan", id_part='name')
+        c.argument('subnet_id', arg_group="Network", options_list=['--subnet'],
+                   help="Resource ID of an existing subnet. If specified, all Codespaces in this plan will be created in this subnet. The subnet must be in the same region as the plan.")
+        c.argument('default_autoshutdown_delay', arg_group="Plan Default", options_list=['--default-suspend-after'],
+                   arg_type=get_enum_type(['5', '30', '120']),
+                   help="Default minutes Codespaces in this plan should suspend after.")
+        c.argument('default_sku_name', arg_group="Plan Default", options_list=['--default-instance-type'], help="Default Instance Type for Codespaces in this plan.")
 
     with self.argument_context('codespace create') as c:
         c.argument('friendly_name', options_list=['--name', '-n'], help='Name of the Codespace.')
-        c.argument('sku_name', options_list=['--instance-type'])
-        c.argument('autoshutdown_delay', options_list=['--suspend-after'], arg_type=get_enum_type(['5', '30', '120']), help="Automatically suspend the inactive Codespace after this many minutes.")
         c.argument('git_repo', arg_group="Git", help="Url of the git repository we'll clone into the Codespace")
         c.argument('git_user_name', arg_group="Git", help="Git username. For example, the output of `git config user.name`")
         c.argument('git_user_email', arg_group="Git", help="Git user email. For example, the output of `git config user.email`")