+++ src/codespaces/azext_codespaces/commands.py
@@ -22,6 +22,7 @@ def load_command_table(self, _):
     with self.command_group('codespace plan', plan_operations, client_factory=cf_codespaces_plan) as g:
         g.custom_command('list', 'list_plans')
         g.custom_command('create', 'create_plan')
+        g.custom_command('update', 'update_plan')
         g.show_command('show', 'get')
         g.command('delete', 'delete')
 
@@ -33,6 +34,7 @@ def load_command_table(self, _):
         g.custom_command('delete', 'delete_codespace', confirmation="Are you sure you want to delete this Codespace?")
         g.custom_command('resume', 'resume_codespace', table_transformer=transform_codespace_item_output)
         g.custom_command('suspend', 'suspend_codespace', table_transformer=transform_codespace_item_output)
+        g.custom_command('update', 'update_codespace', table_transformer=transform_codespace_item_output)
 
     with self.command_group('codespace location', plan_operations, client_factory=cf_codespaces_plan) as g:
         g.custom_command('list', 'list_available_locations', table_transformer=transform_location_list_output)