+++ src/codespaces/azext_codespaces/vendored_sdks/vsonline/vs_online_client.py
@@ -14,7 +14,6 @@ from msrest import Serializer, Deserializer
 from msrestazure import AzureConfiguration
 from .version import VERSION
 from .operations.operations import Operations
-from .operations.account_operations import AccountOperations
 from .operations.plan_operations import PlanOperations
 from . import models
 
@@ -53,15 +52,13 @@ class VSOnlineClientConfiguration(AzureConfiguration):
 
 
 class VSOnlineClient(SDKClient):
-    """Microsoft VS Online REST API version 2019-07-01-preview.
+    """Microsoft VS Online REST API version 2020-05-26-preview.
 
     :ivar config: Configuration for client.
     :vartype config: VSOnlineClientConfiguration
 
     :ivar operations: Operations operations
     :vartype operations: microsoft.vsonline.operations.Operations
-    :ivar account: Account operations
-    :vartype account: microsoft.vsonline.operations.AccountOperations
     :ivar plan: Plan operations
     :vartype plan: microsoft.vsonline.operations.PlanOperations
 
@@ -81,13 +78,11 @@ class VSOnlineClient(SDKClient):
         super(VSOnlineClient, self).__init__(self.config.credentials, self.config)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
-        self.api_version = '2019-07-01-preview'
+        self.api_version = '2020-05-26-preview'
         self._serialize = Serializer(client_models)
         self._deserialize = Deserializer(client_models)
 
         self.operations = Operations(
-            self._client, self.config, self._serialize, self._deserialize)
-        self.account = AccountOperations(
             self._client, self.config, self._serialize, self._deserialize)
         self.plan = PlanOperations(
             self._client, self.config, self._serialize, self._deserialize)