+++ src/codespaces/HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ===============
 
+0.2.0
+++++++
+* Update to 2020-05-26 API version.
+* Plan creation supports subnets.
+* New `az codespace update` command.
+* New `az codespace plan update` command.
+
 0.1.0
 ++++++
 * Initial release.